+++ internals/cuda/matrix_mult.h
@@ -1,7 +1,6 @@
 #ifndef MATRIX_MUL_H
 #define MATRIX_MUL_H
 
-void matrix_multiplication_wrapper(float *a, float *b, float *c, int size);
-void tiled_matrix_multiplication_wrapper(float *a, float *b, float *c, int size);
-
+void matrix_multiplication_wrapper(double *a, double *b, double *c, int size);
+void tiled_matrix_multiplication_wrapper(double *a, double *b, double *c, int size);
 #endif
